requirements.txt
Pillow
scikit-image
scipy
tensorflow==2.0.0
tqdm
matplotlib
numpy
pydot
graphviz